FC mgmt net switch fix
Summary:
A logic to auto select the FC mgmt switch was added for LC. Such logic
shall only be executed on LC. The way to tell if it is a LC or FC is by
using EEPROM. However, the EEPROM is on an i2c bus, which is only
accessible after isolation buffer is enabled. The original diff tries to
access the EEPROM through weutil before the isolation buffer is enabled,
which causes booting hang.

The fix is to move the logic later in setup-gpio.sh after isolation
buffer is enabled.

Test Plan: boot the image on wedge, confirm it boots, no hang anymore.

Reviewers: klahey, aaronmiller, sdasari, simpkins, orib

Reviewed By: orib

Subscribers: net-systems@

Differential Revision: https://phabricator.fb.com/D1896921

Signature: t1:1896921:1425919834:ccf2b0f94a08ad73d48f9ea55d883faae5ed7be6