add an lldp-util tool
Summary:
Add a small utility for receiving LLDP and CDP packets, and printing
information about them.  This is a simple port of some of the FBOSS
code, converted from C++ to C.

Eventually it might be better to install lldpd or lldpad as part of the
image.  (Newer upstream releases of meta-openembedded do have a recipe
for lldpad.)  For now, lldp-util is more convenient since it is a single
file that can easily be copied onto running systems that are still on an
older OpenBMC version.

Test Plan: Tested on a wedge system.

Reviewers: tfang, sdasari, arenzin, klahey

Reviewed By: klahey

Subscribers: ps, doug, net-systems@, exa

Differential Revision: https://phabricator.fb.com/D1909376

Tasks: 4965614

Signature: t1:1909376:1426542911:70af2885e8fb9de38ad656786bac0189ec400f02