fixes to allow us to name GPIO functions
Summary:
There are a bunch of GPIO functions that have moved around onto
various pins over time and across different hardware platforms.
This is an attempt to create a directory of links to specific
GPIO names so that we can try to use the names instead of the
numbers, and write code that works across a variety of platforms
and revisions.

Test Plan:
Look into /tmp/gpionames and see if I correctly added T2_POWER_UP,
which I need for fand

Reviewers: aaronmiller, sdasari, tfang

Reviewed By: tfang

Subscribers: net-systems@

Differential Revision: https://phabricator.fb.com/D2133420

Tasks: 4606061

Signature: t1:2133420:1433920946:0c4521643acbb9489ebee9d4aff28deb19094887