power down, rather than reset, T2 when overheating
Summary:
Previously, we put the T2 into reset rather than powering it off when
the system overheated, and we only triggered all this when the temperature
got pretty high.  Hardware folks prefer that we trigger when the fans
are out for more than a minute.  I have some concerns that this
will result in lots of confused and frustrated data center folks, but
went ahead and set the timeout to around a minute and a half.

Test Plan:
Log into BMC via ssh;  run the following to watch the temperature:

# while /bin/true; do date; sensors fb_panther_plus-i2c-4-40 | tail -3 | head -n 1; sensors tmp75-i2c-3-49 | tail -2 | head -n 1; sleep 10; done

Log in over the serial console:

# tail -f /var/log/messages

Remove the fans, and verify that after a minute or so, the uServer and
T2 are shut down, and that temperatures continue to decrease.

We should probably modify the code to skip the fan timeout checks and
verify that the overheating stuff works as well.

Reviewers: tfang, aaronmiller, sdasari

Reviewed By: sdasari

Subscribers: net-systems@

Differential Revision: https://phabricator.fb.com/D2136911

Tasks: 7319414

Signature: t1:2136911:1433885254:892ed52bc45e9fdc7e51de477117bbcecd00c95c