eth0 mac handling enhancement
Summary:
eth0_mac_fixup.sh is the script to read the MAC from the EEPROM and set
it to the eth0. In the case of an EEPROM is not valid, the eth MAC will
be generated randomlly each time when BMC reboots. This causes different
IP address assigned by the DHCP server as the MAC is changed. It is very
inconvenient to access or upgrade the BMC image.

Test Plan:
Reboot BMC on a unit w/o valid EEPROM. Confirm the same MAC
is used and same IP address is assigned.

Reviewers: klahey, orib, aaronmiller, jmehta, simpkins, sdasari

Reviewed By: sdasari

Subscribers: net-systems@

Differential Revision: https://phabricator.fb.com/D2154987

Signature: t1:2154987:1434388495:3f7f50ea6966e768f40bd7b4e238dd576e3a33c6