Adding the Spatula Wrapper
Summary:
Due to the memory limiations on BMC, we cannot run Chef directly on the BMC.
We have a BMC-proxy running on the uServer to communicate the data from BMC to
ODS and other places. Now we are creating a HTTP endpoint in the BMC-proxy to
serve data to Spatula. Spatula will be a light weight chef style process which
will communicate with the HTTP endpoint and configure the BMC. This will help
us rotate root keys, ssh certs, setting up hostnames and other stuff like ntp.
For ease of updating the Spatula code, we have create a wrapper in the BMC which
would call an endpoint to get the Spatula script. This will then be run by the
wrapper.

The script and all the configs required will be pushed from chef and be made
available with the help of the HTTP endpoint.

Test Plan:
- BMC image build.
- Tested that Spatula was initated at boot.
- Checked the logs for appropriate errors.

And finally:
parrotexplosion

Reviewers: jasmeetbagga, klahey, cooper, dgadling, tfang

Reviewed By: tfang

Subscribers: phild, net-systems@

Differential Revision: https://phabricator.fb.com/D2237669

Tasks: 5933047

Signature: t1:2237669:1438368611:651b6c1f40c11dc74117a6ab0a521777325d65e2