remove Panther+ support
Summary:
As we are switching to support COMe only, remove support for Panther+ to
simplify the code.

Also removed the uServer power monitor script as it is not applicable to
COMe anymore.

Test Plan: Load the image on a wedge100 w/ COMe.

Reviewers: klahey, aaronmiller, jmehta, simpkins, vineelak, sdasari

Reviewed By: sdasari

Subscribers: net-systems-diffs@, yogeshwer

Differential Revision: https://phabricator.fb.com/D2469413

Signature: t1:2469413:1442966183:3cdcc34aab287533fdba5aaf8d7cfafd3d8b1789