Add a timeout for popen so if the endpoint 'sensors' gets stuck , process can be terminated.
Summary: Fixes for t8395648. Add timeout for popen so if the RESTful server gets stuck , process can be terminated.

Test Plan:
Added logs to output to a file and verified that the process gets killed when timeout occurs.
Tested on a wedge by adding a sleep() after popen and before we read the data. When the api timesout , process is killed.
https://phabricator.fb.com/P20029811

Reviewers: adhipgupta, simpkins, tfang

Reviewed By: tfang

Subscribers: net-systems-diffs@

Differential Revision: https://phabricator.fb.com/D2457672

Tasks: 8395648

Signature: t1:2457672:1444069908:b205eddede5cb4f3fb54202f71d8a0348635d036