Tool changes to fix session end, BIOS entry, help
Summary:
Diff contains a few changes including:
- To brak into BIOS from sol.sh session, using fn+DEL
is not being handled by the tool because it was receiving
in-correct bytes for the action. For now hacking to log the
event and dropping the command instead of closing the client session.
Alternately ESC key can be using , hence the action is not blocked.
- Improvements to client session termination.
- Changes to help message to use log file for now and not the tool to
view the log.

Test Plan: verified on lab switch

Reviewers: tfang

Reviewed By: tfang

Subscribers: net-systems-diffs@

Differential Revision: https://phabricator.intern.facebook.com/D3618833

Tasks: 9288137

Signature: t1:3618833:1469577519:70aade64179fc2b77962f1194ffb4cc865bd5b1f