Change FBTP's bootm arguments to boot a FIT
Summary:
Part 3 (and final) of the initial verified-boot series for FBTP.
This changes the `bootm` command to point to a FIT.
We place the FIT at the same SPI address as the kernel (or rather, where the kernel used to be).

Since `bootm` is no longer booting a kernel + rootfs, only 1 parameter is needed.
This command will 'discovery' the image type using the FIT/FDT magic.

Test Plan:
`bitbake fbtp-image`
Boot on a Quanta FBTP in the lab.
This is part 3, together with part 1/2 this should load and boot a kernel + rootfs.

Reviewers: tfang, jmehta, sdasari

Reviewed By: sdasari

Subscribers: net-systems-diffs@, mcgrew

Differential Revision: https://phabricator.intern.facebook.com/D3708516

Tasks: 12663957

Signature: t1:3708516:1470976980:38cf5e8a98933e9eed697fe50e5862c41535838a