[fbtp] Allow couple of thresholds to be dynamically updated
Summary:
The cpu0/cpu1's TjMAX indicates threshold value for cpu temperature and is unique to each
processor. Since these values are available only when the cpu is powered ON, this diff allows
the thresholds to be initialized dynamically when the system is powered ON.

Test Plan: Build and test on fbtp platform.

Reviewers: vineelak, kapilhali, mattconolly, nicolereid, reed, jmehta, siumingtong, jna, jfern, aaronmiller, tfang

Reviewed By: tfang

Subscribers: net-systems-diffs@

Differential Revision: https://phabricator.intern.facebook.com/D3763161

Signature: t1:3763161:1472073072:bc9f9a108755cc1747ffbd35f235c526ac0f39a4

fbshipit-source-id: bad6313