[u-boot] Towards a multi-SPI boot
Summary:
This adds two new defines for controlling a DUAL_SPI or "ROM boot" for U-boot.
It should only affect AST2500-based boards using helium. It should be an opt-in and every board to build a ROM(SPL) or non-ROM U-boot.

The major changes occur in the 3-linked board definition includes:
```
./BOARD_NAME.h
./facebook_common.h
./ast2500_common.h
```

The changes include two defines:
```
CONFIG_FACEBOOK_DUAL_SPI
```
This triggers a SPL build and defines most of the U-boot memory locations to offset from `0x3000:0000` the location of the AST FMC MMIO for SPI1.

```
CONFIG_FACEBOOK_VERIFIED_BOOT
```
This enables U-Boot signature verification, for the actual U-boot, from the SPL build.
Thus it is possible to enable a DUAL-SPI build without enforcing a verified boot.

The next set of changes are to the board build:
AST-G5's reset/init code was copied into a barebones `-spl` version. This executes 100% before code-relocation.
These changes are in C and implement the U-boot FIT parsing, multi-key verification, and enforcement.

The last set of changes are to the G5 platform initialization code.
This breaks the code into 2 sections, with some of the initialization occuring in both.

The `CONFIG_SPL_BUILD` defines code that executes in the first SPI, the eventual ROM.
Anything not gated in this include is included in BOTH. So for code like the `init_dram` DRAM calibration, we invert the check.

Test Plan: Not yet, this is designed for patching into a Poky build and further testing the `verified-boot` recipe.

Reviewers: jmehta, srihan, tfang, sdasari

Reviewed By: sdasari

Subscribers: thor, net-systems-diffs@, mcgrew

Differential Revision: https://phabricator.intern.facebook.com/D3750631

Signature: t1:3750631:1471883940:88d19fe95b37f2f539c3144236650a0f2250d409

fbshipit-source-id: 02a2ecf