[u-boot] Reset SPI0 CS1 state in SPL boot
Summary:
The AST2500 SPI1 (FMC CS1) is a SPI0.1.

For the SPI0.0 (`0x2000:0000`) the FMC will detect the chip status/register and set
CS0 status (see `0x1E62:0000`). This is not the case for SPI0.1.

We have two options within the ROM/SPL:
1. Reset all of the SPI part settings/modes.
2. Attempt to detect the modes and set the AST FMC registers appropriatly.

This is choosing 1 such that we do not miss a setting or pair of setting - register.

Test Plan:
Build: `fbtp_defconfig` with the `CONFIG_SPL=y` uncommented.
Flash and soft-reset the AST.

Reviewers: tfang, jmehta, sdasari

Reviewed By: sdasari

Subscribers: net-systems-diffs@, mcgrew, thor

Differential Revision: https://phabricator.intern.facebook.com/D3795424

Signature: t1:3795424:1472660305:5516d7a4c0661abf9b3fe46a9a0ed88d61b71474

fbshipit-source-id: 5a4739b