[Linux] Update MTD partitions: rename kernel to u-boot and expand rootfs
Summary:
This is a **BREAKING CHANGE** for some tooling (but not really).
The `kernel` and `rootfs` partitions have been re-assigned to `u-boot` and `fit` respectively.

With dual SPIs, both have the same content written.
The `SPI0.0` is write protected with `WP#` (and proper chip registers) and detects as much thus executing `SPI0.1`'s u-boot after running it's ROM portion.
The `SPI0.1` has a "dead area" called ROM that is useless and never executed since it's really RW.
To accomodate this layout compatibility the partitions are pushed forward.

The now-combined `kernel` and `rootfs` are called `fit` and the area has been expanded to `27.125M`.

Test Plan: `bitbake fbtp` and observe the regions in the written `flash` deploy image.

Reviewers: nicolereid, siumingtong, sdasari, tfang

Reviewed By: sdasari, tfang

Subscribers: net-systems-diffs@, mcgrew, thor

Differential Revision: https://phabricator.intern.facebook.com/D4144936

Signature: t1:4144936:1480721208:f2c2ad4b7d9aba309ac63793ab8609ea94df7e92

fbshipit-source-id: 805b4af