verified-boot: Clean up Yocto build logic
Summary:
This is a large clean up of the verified-boot logic within `meta-aspeed`'s `u-boot` recipe.

1. The removes the `rom-boot` IMAGE_FEATURE that was used during testing.
There was a period where one could perform a `rom-boot` but avoid the verified-boot flow.
This is no longer the case, you should use the software verify method for testing.
By default `verify=no` so the verified-boot flow is exercised, but not enforced.

2. U-Boot compilations not occur within a `./default` relative folder.
Check to make sure all `u-boot.bin` and `u-boot.elf` and `u-boot-spl.bin` binaries are
using a `./default/` directory within `do_install` and `do_deploy`.
This allows for a secondary U-Boot build within `./recovery`.
The logic will flip a configuration option for the Facebook board that enables the
recovery flow and on-ROM (`0x2000:0000`) in-place execution.

3. Move all hash algorithms to `sha256` via the `VERIFIED_BOOT_HASH_ALGORITHM` variable.
The `VERIFIED_BOOT_ALGORITHM` is the signing algorithm, set to `sha256,rsa4096`.

4. Remove older and unused FIT metadata.

Test Plan:
1. Build without the `verified-boot` image feature and assure:
* A single U-Boot is built.
* U-Boot addresses and the kernel/rootfs FIT are relative to `0x2000:0000`.

2. Build with `verified-boot`:
* Make sure `verify=yes` is NOT set in the environment.
* Use 2 SPIs and make sure execution continues into `0x2800:0000`.
* Set `force_recovery=yes` and assure U-Boot executes from `0x2000:0000` and runs the DHCP/TFTP flow.
* Set `verify=yes` and assure the same recovery flow executes.

Reviewers: tfang, sdasari

Reviewed By: sdasari

Subscribers: net-systems-diffs@, dguerri, mcgrew, thor

Differential Revision: https://phabricator.intern.facebook.com/D4392062

Signature: t1:4392062:1483986092:9d4f78a1e1d6907c7c1571040d42248644419925

fbshipit-source-id: 5373175