u-boot: Add an upper bound to recovery retries
Summary:
Apply an 'impossible' upper bound to recovery retries: 25.
If the values in the word is greater than 25 the assumption is random unsanitized data.

This content is provided to be helpful in debugging, but we want it sane and consistent.

Test Plan: Run `vbs.sh` in Linux, check that the value is not something like `0xE5`.

Reviewers: tfang, sdasari

Reviewed By: sdasari

Subscribers: mcgrew, thor

Differential Revision: https://phabricator.intern.facebook.com/D4500695

Signature: t1:4500695:1486073868:420e868dab89e141c64efc9f72520ea8a89acac2

fbshipit-source-id: 59e9ec9