u-boot: Abstract WDT method and apply 60s timeout to SPL
Summary:
There are now three places in U-Boot where the ASPEED WDT is used.
Because of this we need to abstract the call to two inputs: `timeout` and `mask`.
The timeout is set in micro-seconds with a normalized 1M clock added to the existing 300s timeout.

The CPU reset method in U-Boot uses a `0x10` for the clock.

We introduce a new reset call within the SPL of `60s`, this gives the ROM/SPL 60 seconds to boot U-Boot.

Test Plan:
1. Build, run `fbtp`.
2. Assure that U-Boot is executed, interrupt the Linux boot and wait.
3. Assure 300s have passed before the reset.

Alternatively, place a hang before U-Boot is executed and expect a 60s reset back into ROM.

Reviewers: tfang, sdasari

Reviewed By: sdasari

Subscribers: mcgrew, thor

Differential Revision: https://phabricator.intern.facebook.com/D4549264

Signature: t1:4549264:1487014646:d662bb8e834e7f0d92d3517b311251d96b2f85d4

fbshipit-source-id: 7741d29