Install /mtd/datax partition for single-SPI storage
Summary:
For the FBTP, with verified-boot enabled, and a single SPI populated, there is no `/dev/mtdblock4` (or `data0` partition).
This creates a MTD partition called `datax` that exists in the location of `data0` on the RO SPI.
This is helpful while transitioning from non-verified-boot to a verified-boot when only one SPI is used.

Test Plan:
1. Build and boot FBTP without a second SPI.
2. `mount`, the MTD device for `datax` should be mounted at `/mnt/data/`.

Reviewers: sdasari, tfang

Reviewed By: sdasari

Subscribers: net-systems-diffs@fb.com, mcgrew, thor

Differential Revision: https://phabricator.intern.facebook.com/D4563383

Signature: t1:4563383:1487184337:cac74d3063d505cbb9171367cf83740ffcceee67

fbshipit-source-id: e79ea2a