Fix fand for 4.1 kernel
Summary:
In 4.1 kernel, the temperature sysfs file is moved to
device/hwmon/hwmon???. The diff adds the logic to find out '???' and
read the temperature in either directly in device or
device/hwmon/hwmon???

Test Plan:
Boot the image with 4.1 kernel on wedge. Before the fix, fand
exits because it cannot read any temperatures. With the fix, fand is
able to read the temperature and manage fans.

Reviewed By: sdasari73

fbshipit-source-id: 2013933