Do not clear the boot-order flag when boot-order is requested
Summary:
Do not clear boot-order when requested.
This fixes the issue when ipmi-tool is used to set the boot order
and immediately query the boot order to verify if it is set only
to discover that 'get' ends up clearing it.

Test Plan: build, boot-up and ensure host boots up.

Reviewed By: sdasari73

fbshipit-source-id: 30ab114