lightning: Rounding current sensor value to the nearest hundredth
Summary:
Rounding current sensor value to the nearest hundredth before checking the threshold and stored the value in sensor cache.
To fix The sensor value which is used to threshold checking is different from which is showing to user by sensor-util.
(This fix is only for lightning.)
Closes https://github.com/facebook/openbmc.wiwynn.lightning/pull/182

Test Plan: Test on lightning platform pass.

Reviewed By: jmehta

fbshipit-source-id: 144400d