verified-boot: tests: Add Dockerfile for end-2-end tests
Summary:
This adds a `Dockerfile` and test harness for CI of the verified-boot flow.
This is important for the open source project, and will be used with TravisCI to make sure our ShipIT/sync does not effect the verified-boot security.

This does have failure tests, so the `Dockerfile` is rather complex.

Test Plan:
On your laptop, or somewhere not-prod:
```
docker build .
```
If it exits 0, you're set to go.

Reviewed By: tfangit

fbshipit-source-id: ad3fb1b