tpm: Add TPM-based rollback protection for verified-boot
Summary:
This adds rollback protection using the TPM to OpenBMC verified-boot.

This is not a complete feature as it is missing two components (gLock, ownership).
Including this provisioning code will not cause any PERM changes to the TPM state.

Test Plan: This requires several rounds of testing. Please stand by.

Reviewed By: sdasari73

fbshipit-source-id: f3419d7