vboot: Refactor SPI flash code to detect FMC SPI0.0 status
Summary:
This commit refactors the `flash-spl` implementation to be able to operate on FMC SPI0.0.
Issuing commands to the SPI and FMC controller requires us to run code from SRAM (heap).
Running code in-place will always reset the controller state when fetching and executing.

This update logic performs two write-status-register operations on each SPI, then resets
the controller state for each chip select (returning them to 32bit addressing mode).

The two rounds of write-status-register checks the HW protections.
If the `WP#` pin is active low then a write of `1` to `SRWD` will "lock" updates to the
status register. In this first round we write `1` to the entire set of block protections.

If we can re-write the status register then the `WP#` is not active low and hardware
protections are not enforced. To remove abiguity about the protections the block-protections
are removed in the second round if it succeeds.

This should also report the status of the first write operation back to the caller.
The caller can use this status to determine if the second SPI is populated.

Test Plan:
1. Run the SPL without a FMC SPI0.1
2. Run the SPL with a FMC SPI0.1 (check HW enforcing via `vbs` in U-boot`).
3. Run the SPI with a FMC SPI0.1 and with WP# low, the hardware enforcing should be set.

Reviewed By: sdasari73

fbshipit-source-id: 1770971