Bryce Canyon: Fix the bug of SCCB missing could not be detected on IOMB
Summary:
Due to SCC location is solid on FDPB.
GPIO expander P06 is for SCC A, P07 is for SCC B.
Our original design that let SCC B present status could not be detected on side B.
So we could not know the SCC B was missing.
We modify the feature to confirm that the type 5 IOM is in which side.
Then get the SCC present status from GPIO expander to check SCC is missing or not.

Closes Https: //github.com/facebook/openbmc.wiwynn.lightning/pull/258

Test Plan:
It works on Bryce Canyon EVT system.

commit: b0b877cd348473537c24dc48cf74efd8f2cab448

Reviewed By: jmehta

fbshipit-source-id: 222d5a0