fbttn: Add NIC temperature workaround
Summary:
Add the workaround for the NIC temperature 0xFF value.
This workaround is merely to unblock the DVT build candidate.

Closes Https: //github.com/facebook/openbmc.wiwynn.lightning/pull/276

Test Plan:
It works on Bryce Canyon EVT system.
Because the latest sensord.c which let sensors reading keep to NA, so I back to master ac9b79a then use it's sensord.c to verify.

Reviewed By: jmehta

fbshipit-source-id: 3dd6786