fbttn: Modify the initial value of GPIOA3 (SYS_PWR_LED).
Summary:
Since HW pulled up resistor of the GPIO in DPB, the LED status keeped to ON.
If BMC set the GPIO initial value to low, it will let LED from ON to OFF.
The LED will be set to ON after BMC is ready and running front-paneld.
So the LED status was ON->OFF->ON.
It may cause user confusion.
So we modify the initial value of GPIOA3 (SYS_PWR_LED) to high.
Keep the LED always ON until BMC is ready.

Closes Https: //github.com/facebook/openbmc.wiwynn.lightning/pull/287

Test Plan: It works on Bryce Canyon EVT system.

Reviewed By: jmehta

fbshipit-source-id: f3a287c