fbttn: Fix power-off server 12V not work issue
Summary:
Due to we added the feature that BMC Power on server 12V after detected SCC or MonoLake be pused in.
It cause power-off server 12V is not work.
Because we will auto power-on server 12V when Mono Lake and SCC are present.
So add a flag to confirm MonoLake or SCC is missing then check the flag to decide power-on server 12V.

Closes Https: //github.com/facebook/openbmc.wiwynn.lightning/pull/291

Test Plan: It works on Bryce Canyon EVT system.

Reviewed By: jmehta

fbshipit-source-id: f71d5f6