vboot: Add create-tests script to generate unit-tests
Summary:
This commit introduces a `tests/verified-boot/create-tests.sh` script.
This script should be run after building a `flash-MACHINE` flash and is intended
to create several 'sets' (CS0 + CS1) of flashes.

Each test set is labled with `N.MM` where `N` and `MM` are the error type and
error code expected after flashing and booting without HW or SW enforcement.

If using HW or SW enforcement the error code should be the same, but for each
test other than the `0.00` test, the recovery U-Boot/flash will be used.

Test Plan:
1. Run this added script.
2. Flash each set of flashes and verify the result of `vbs` (U-Boot) or
`vboot-util` (Linux) matches the `N.MM` in the flashed file.

Reviewed By: sdasari73

fbshipit-source-id: 6073eb1