vboot: Check U-Boot hash and improve non-enforce flow
Summary:
This applies regression fixes found with the unit tests from: D4945037.
The largest change includes manually checking the hash of the U-Boot content.

When the flow was switched from an image-based signature verification to conf-based,
it was assumed the configuration signature API call would check the included images.
This is not true, it only verifies the FDT content of the node-list.

The "non-enforce flow" is improved with some changes to the flat-tree of check and fails.
Some actions taken after the machine passes into a non-enforce vboot failure mode make the
subsequent errors and output confusing.

E.g.: We should not change the signature store to the subordinate store when the subordinate
fails verification.

Test Plan:
1. Build `bitbake fbtp-image`.
2. Run the `create-tests.sh` script and verify each state.

Reviewed By: sdasari73

fbshipit-source-id: 60c9358