vboot: Enable TPM and TPM reset for FBTP
Summary:
The current FBTP machines use `GPIOAB2` to reset the TPM when the CPU is reset.
The `CONFIG_ASPEED_TPM` will turn on the TPM provisioning and fallback protection checks in the SPL.

Test Plan:
After running a build for FBTP, use the `create-tests.sh` script in `./tests/verified-boot`.
Flash and reboot several times, you should continue to get `0:0`.

Resign the CS1 image:
```
../meta-openbmc/tests/verified-boot/signing/fit-sign \
  --kek /tmp/outputs/kek/kek.dtb \
  --signed-subordinate /tmp/outputs/subordinate/subordinate.dtb.signed \
  --keydir /tmp/outputs/subordinate \
  --mkimage ./tmp/sysroots/x86_64-linux/usr/bin/mkimage \
  ./tmp/deploy/images/fbtp/flash-fbtp /tmp/outputs/flash-fbtp-resign
``

Do it again but save this second resign as `flash-fbtp-resign2`.

Flash `flash-fbtp-resign2` to CS1 (`--bmc`) and reboot.
Flash `flash-fbtp-resign` to CS1 (`--bmc`) and reboot.

The rollback protection should report an error with `vboot-util`.

Reviewed By: amithash

Differential Revision: D5018002

fbshipit-source-id: 767b1b3