Increase sleep time after enable baterry sense pin to 10ms
Summary:
We built a test FW to pull another pin to low after msleep(1), and we captured the signal with scope, we found the interval time between sense enable and another pin low, they were around 760us to almost 10ms in several times, it showed the msleep(1) could be less than 1ms. After we changed msleep(1) to msleep(10), it interval increased to 10 ~ 20ms.

The msleep() in pal.c call the function nanosleep()

https://linux.die.net/man/2/nanosleep
If the interval specified in req is not an exact multiple of the granularity underlying clock (see time(7)), then the interval will be rounded up to the next multiple. Furthermore, after the sleep completes, there may still be a delay before the CPU becomes free to once again execute the calling thread.

BMC's HZ is 100 and the jiffy is 10ms, according Linux man page, the nanosleep should round up the 1ms to 10ms, and with a delay after sleep completed, it should sleep 10 ~ 20ms, somehow it only sleep 0 ~ 10 ms in BMC.

This PR is for eTrack#201144 and #201143 in fbtp-v1.7 (1.7 without retry)

Reviewed By: sdasari73

Differential Revision: D5045581

fbshipit-source-id: 02a319c