sensord: Fix issue where thresholds are not tested on the last retry
Summary: On the last retry, we break out of the loop before testing if the threshold is met. This change ensures that every sensor read is followed by a test of the threshold.

Test Plan: Sanity test on fbtp. Also the change is being sent to ODMs for further validation.

Reviewed By: sdasari73

Differential Revision: D5046442

fbshipit-source-id: 48720b8