vboot: Add verified-boot tests for TPM and OS
Summary:
This adds several new tests for verified-boot.
These test cases stress the OS signing and TPM rollback protection.

Test Plan:
Run the `create-tests.sh` and the `run-tests.py` QEMU launcher.
For each of the flashes created the `vbs` or `vboot-util` should return the status relative to the file name.

Reviewed By: vineelasmile

Differential Revision: D5051189

fbshipit-source-id: 14f6096