vboot: Add TPM PCR extending to SPL
Summary:
This brings fast PCR extending to the U-Boot SPL for the AST 2500.
We define some basic use of the PCRs (0-7) for the SPL, subordinate keys, U-Boot, and the environment.

SPL: This is the ROM, the region that contains the KEK and checks the subordinate keys and U-Boot.
Subordinate keys: These are the ODM and 'online' signing keys.
U-Boot: The second-stage firmware.
Environment: Assume these variables are immediately changing the U-Boot execution.

Test Plan: Build and run, boot into linux, use `tpm-tools` or `tool4tpm` to inspect PCRs 0-3.

Reviewed By: sdasari73

Differential Revision: D5226639

fbshipit-source-id: 1c6ae69