vboot: Simplify PKCS11 engine use and do not set key-name
Summary:
This is a follow up to the diff bringing upstream PKCS11 support to `mkimage`.
The goal was to isolated the upstream changes from our own.

This does NOT set the `key-name-hint` to the URL.
The key hint is expected to be set before hand within the input DTS.

Test Plan:
Build and using `create-tests.sh`, and boot.
The verified key name should be the name of the input folder (`odm0`, or `odm1`).
It should NOT be `pkcs11:....`.

Reviewed By: sdasari73

Differential Revision: D5234999

fbshipit-source-id: 53bd7a3