vboot: Reset once if TPM returned invalid postinit
Summary:
This adds a single reset retry if the SPL receives an invalid-postinit (38) from the TPM.
An invalid postinit will occur if the BMC is reset without resetting the TPM.

The TPM is reset when the WDT1 EXTRST is used to reset the BMC. Any other forms of reset
may not configure the WDT to pulse EXTRST or pulse it with the right configuration.

The SPL will, so a single reset should be adequate.

Test Plan:
1. Flash, reboot, and run `vboot-util`.
2. Use an external reset, other than power-failure, and run `vboot-util`.

Reviewed By: amithash

Differential Revision: D5273816

fbshipit-source-id: f389540