fbtp: Add machine configuration string support
Summary:
Interpret the machine configuration received from the BIOS and store a single string
to represent the machine configuration. This should be useful for other modules to
modify behavior based on the machine configuration used.

Test Plan:
Load on FBTP and ensure nothing abnormal is seen. BIOS is yet to provide
           this command, so this can be tested only in unit-test fashion.

Reviewed By: sdasari73

Differential Revision: D5342055

fbshipit-source-id: af1bc36