wdt: Use a SoC reset for WTD1 in U-Boot
Summary: Peform a SoC reset within U-Boot, right now we are doing a full-reset.

Test Plan:
This needs a bit of testing.

1. Flash to a FBTP and interrupt U-Boot's flow before loading the kernel.
The expectation is the machine will reset within 2mins.

2. Allow a normal boot into Linux.
The expectation is the machine will not reset.

3. Reboot from U-Boot using `reset`.
The expectation is to boot into recovery, because the machine does not know the CLI was used.

4. Reboot from Linux using `reboot`.
The expectation is to reboot normally back into Linux.

Reviewed By: sdasari73

Differential Revision: D5315374

fbshipit-source-id: 5ce96f6