vboot: Reset handoff and WDT1 when using U-Boot console
Summary:
This will stop the WDT1's 2m timer when using the U-Boot console.
If someone interrupts the U-Boot boot, and drops to the console, they do not expect the machine to resume.

The ROM handoff marker is reset to tell the SPL that there was not an error.

Test Plan:
1. Boot normally, you should arrive at Linux on CS1.
2. Boot and interrupt U-Boot, then `reset` and boot normally, you should arrive at CS1.
2. Boot and interrupt U-Boot, delay for >2m, you should still be at the U-Boot prompt.

Reviewed By: sdasari73

Differential Revision: D5319010

fbshipit-source-id: fefd372