uboot: Enable correct I2C bus based on AST address
Summary:
The I2C uclass drivers use their sequence in the embedded device tree to identity the I2C bus.
In U-Boot, the SPL does not relocate so any global data structures are static.

This fixes an error where the `i2c7` alias is not identified as bus 7, but rather 1.
This is OK except for the I2C bus enable SCU method, which needs the correct bus number.

We forget the sequence, or bus property, and use the address to identify the AST bus.

Test Plan: Run `vboot-util` after a `power-util sled-cycle`.

Reviewed By: amithash

Differential Revision: D5369565

fbshipit-source-id: 0b27f93