RestAPI - Code conversion to Py3 + Transition from CherryPy-bottle framework to asyncio-aiohttp
Summary:
The RestAPI works with asyncio-aiohttp framework. I have attached screenshots for all the endpoints except for a POST request on psu_update. A GET request to rest_gpios.py doesn't produce any output. This is because of the files that are missing in the this directory : ('/sys/class/gpio/gpio%d/value' % gpio, 'r'). Otherwise, the code runs fine. This was tested by importing all the necessary python 3 packages directly onto the BMC rather than building it and flashing a new image every time. So, for this to work, I would be adding a few more recipes to the helium build so that these packages are available on the BMC (this would be my next diff).

Note: I don't intend to land these diffs until the completion of the entire project. This diff is merely for code review purpose.

Test Plan:
Screenshots.
{F67721137}

{F67721140}

{F67721144}

{F67721155}

{F67721167}

{F67721168}

{F67721169}

Differential Revision: D5260394

fbshipit-source-id: b45565d