vboot: Disable preemptive FMC SPI SRWD usage
Summary:
This disables the usage of the SRWD (status register write disable) bit for the FMC SPIs.
This means boards may ship with WP# for CS1 and CS0 and no R/O effect will take place.

The protection bits can be easily removed as the SR values are not locked until the SRWD is set.
If someone/something does set this bit then the protections on that BMC will hold until next cold reset.

Test Plan:
0. Setup, choose an FBTP with FMCSPICS0/FMCSPICS1 held low.
1. Write this U-Boot/SPL to both ROM and BMC and reboot.
2. Try to write again, it will succeed.

Reviewed By: jmehta

Differential Revision: D5488028

fbshipit-source-id: 6ad7e99