vboot: Fail verification for kernel if keys are unavailable
Summary:
This diff will enforce verified boot for the rootfs and kernel if all preliminary checks have passed.

The verified-boot for the rootfs and kernel is taken mostly from upstream U-Boot.
There is no concept of an check but not enforce, either you check and enforce or disable verified-boot logic all together.

To compensate we first check if the vboot error == `VBS_SUCCESS`.
If this is true then almost all vboot checks have passed (except for the OS verification).

This means if the OS fails verification the recovery flow will be taken.

This diff also improves the failure case for the OS verification.
If OS verification had failed, U-Boot would stop for 5mins at the prompt then reset into the recovery flow.
This will now immediately boot the recovery OS, while preserving the failure codes (type=6 for OS).

Test Plan:
We introduce 4 new tests for the OS failure.

1. Boot a signed OS and expect hash/sha256 checks.

2. Boot a signed OS with a missing subordinate key (edit the hint name).
Expect a failure in this case and to boot into the recovery OS with error [6, 60].

3. Boot a signed OS with a corrupted region `0xE:1000` (as random).
Expect the same as 2, to boot into the recovery OS with error [6, 60].

4. Boot an unsigned OS, without subordinate keys, or another failed vboot component.
Expect no hashing for the OS, and to boot into the normal OS.

Reviewed By: benwei13

Differential Revision: D5594506

fbshipit-source-id: 28c70d5