vboot: Increase max SPL FIT to 5120 bytes
Summary:
To support backup KEK keys in the SPL FIT and multiple placements of TPMs,
the max size of the FIT should be increased to 5k. This allows for 79kB for SPL
space, which several kB larger than the current.

Test Plan: Use `fit-sign` to sign a KEK DTB that contains multiple keys for an FBTP machine.

Reviewed By: benwei13

Differential Revision: D5620779

fbshipit-source-id: 7a954d1