vboot: Set TB bit for MX WB and DG SPI NOR flash
Summary:
The goal is to set the Top/Bottom bit for all supported SPI NOR flash chips.

On the MXIC childs this is in the 'configuration' register, so a few more methods are needed.
We detect the SPI chip version based on the first byte of the SPI ID.
The assumption is that each vendor is using their own T/B implementation and that is standard within all of their chips.

Test Plan:
1. Write this and boot, test a successful boot. This changes SPL code.
2. Enable the `CONFIG_ASPEED_FMC_SPI_LOCK` on the test machines with a WP# pull down.
3. Try to write to `0x0` on `/dev/mtd5` (CS1). If you cannot, then success.

Reviewed By: amithash

Differential Revision: D5600282

fbshipit-source-id: 3d55105