Add support for GCC 6
Summary:
This adds support for GCC 6.0 used by Yocto Piro. It just copies over the v5 file as v6.
This is a temp. workaround since this issue is fixed in later kernel releases. But till we upgrade, this should be a sufficient workaround.

Test Plan: Build helium on Yocto/OE Piro. Obviously wont fully build (yet), but at least the kernel should build.

Reviewed By: benwei13

Differential Revision: D5704465

fbshipit-source-id: 66856c9